% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{add_slice}
\alias{add_slice}
\title{helper function for adding a slice to the plot, allowing for additional panel attributes to be passed through}
\usage{
add_slice(
  coordinate = NULL,
  title = NULL,
  bg_color = NULL,
  text_color = NULL,
  border_color = NULL,
  border_size = NULL,
  xlab = NULL,
  ylab = NULL,
  theme_custom = NULL
)
}
\description{
helper function for adding a slice to the plot, allowing for additional panel attributes to be passed through
}
\details{
I'm still unclear about whether it would be better to unify with this annotate_panels in some way.
}
