% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{add_slices}
\alias{add_slices}
\title{Add slices to a ggb object}
\usage{
add_slices(slices = NULL)
}
\arguments{
\item{slices}{a character vector of slices to be added to the ggb plot object}
}
\value{
a ggb object with the relevant slices and an action of 'add_slices'
}
\description{
Add slices to a ggb object
}
