% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{annotate_panels}
\alias{annotate_panels}
\title{Adds custom annotations to one or more panels on the ggbrain plot}
\usage{
annotate_panels(panel_data)
}
\arguments{
\item{panel_data}{a list or data.frame containing annotations to be added}
}
\description{
Adds custom annotations to one or more panels on the ggbrain plot
}
