% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_slices.R
\name{apply_definition_3d}
\alias{apply_definition_3d}
\title{Apply a definition/contrast string to a 3D image volume}
\usage{
apply_definition_3d(img_obj, definition)
}
\arguments{
\item{img_obj}{A ggbrain_images object}

\item{definition}{A definition string, e.g., a simple image name or a filtered expression}
}
\value{
A 3D array with the contrast applied (masked/thresholded volume)
}
\description{
Apply a definition/contrast string to a 3D image volume
}
\keyword{internal}
