% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_neighbors}
\alias{count_neighbors}
\title{This function counts the number of neighboring/touching pixels in a 2D binary image}
\arguments{
\item{im}{A boolean matrix representing a binary image}

\item{diagonal}{Whether to count diagonal elements as valid neighbors}
}
\value{
A matrix of the same size as \code{im} containing the number of neighboring pixels
}
\description{
This function counts the number of neighboring/touching pixels in a 2D binary image
}
\details{
This is an internal function used by geom_outline to clean up outlines
}
\author{
Michael Hallquist
}
