% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_from_edge}
\alias{fill_from_edge}
\title{This function finds holes by flood filling TRUE into a 2D binary image, starting from the edge}
\arguments{
\item{im}{A boolean matrix representing a binary image}

\item{nedges}{An integer specifying how many starting points along the edge to use for
filling TRUE. The starts are northwest (1), southwest (2), southeast (3), and northeast (4).
The compute time increases with the number of starts.}
}
\value{
A matrix of the same size as \code{im} containing the number of neighboring pixels
}
\description{
This function finds holes by flood filling TRUE into a 2D binary image, starting from the edge
}
\details{
This is an internal function used by geom_outline to clean up outlines
}
\author{
Michael Hallquist
}
