% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flood_fill}
\alias{flood_fill}
\title{This function flood fills a binary image with TRUE for any value of FALSE}
\arguments{
\item{im}{A boolean matrix reference representing a binary image}

\item{x}{the starting x position for fill}

\item{y}{the starting y position for fill}

\item{r}{the number of rows in im}

\item{c}{the number of columns in im}
}
\value{
Nothing. The matrix \code{im} is modified in place (by reference)
}
\description{
This function flood fills a binary image with TRUE for any value of FALSE
}
\details{
This is an internal function used by geom_outline to clean up outlines
}
\author{
Michael Hallquist
}
