% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{geom_brain}
\alias{geom_brain}
\title{Helper function to add a contrast layer to the plot. This is a thin wrapper around ggbrain_layer}
\usage{
geom_brain(...)
}
\arguments{
\item{...}{all parameters are passed through to ggbrain_layer$new() to create a new layer object}
}
\value{
a ggb object populated with the ggb_layer and the action of 'add_layer'
}
\description{
Helper function to add a contrast layer to the plot. This is a thin wrapper around ggbrain_layer
}
\details{
Note that the color_scale and limits must be specified at the time of the geom_contrast creation
in order for them to be mapped properly within ggplot. Because we overlay many raster layers in a ggplot
object that all use the fill aesthetic mapping, it becomes hard to map the color scales after the layer is
created using the typical + scale_fill_X syntax, and similarly for scale limits.
}
