% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{geom_outline}
\alias{geom_outline}
\title{Variant of geom_brain that only draws outlines, but does not fill them with stats}
\usage{
geom_outline(...)
}
\description{
Variant of geom_brain that only draws outlines, but does not fill them with stats
}
