% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{geom_region_label_repel}
\alias{geom_region_label_repel}
\title{Variant of geom_label_repel used for plotting region labels on slices with separation from other labels}
\usage{
geom_region_label_repel(image, label_column = "label", ...)
}
\arguments{
\item{image}{The name of the image within the underlying ggbrain_slices object that contains the labeled data positions}

\item{label_column}{The column name name for the labels to use within the slice data}

\item{...}{All other parameters passed through to geom_label_repel}
}
\value{
a \code{ggb} object with the relevant ggbrain_label field and an action of "add_region_labels"
}
\description{
Variant of geom_label_repel used for plotting region labels on slices with separation from other labels
}
