% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{ggb}
\alias{ggb}
\title{Generic R6 base class that is used to support + semantics}
\description{
Generic R6 base class that is used to support + semantics

Generic R6 base class that is used to support + semantics
}
\details{
this object becomes a simple storage class that contains all relevant objects
(e.g., ggbrain_images) required to generate a brain plot
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ggb_images}}{ggbrain_images object for this plot}

\item{\code{ggb_image_labels}}{a named list of data.frames that label corresponding images}

\item{\code{ggb_slices}}{character string of slices to extract}

\item{\code{ggb_layers}}{a list of ggbrain_layer objects containing the bottom-to-top layers to be plotted}

\item{\code{ggb_plot}}{a ggbrain_plot object containing the specification of the plot}

\item{\code{ggb_annotations}}{a list of annotation objects}

\item{\code{ggb_region_labels}}{a list of ggbrain_label objects to be added as text to label regions}

\item{\code{action}}{what should this ggb object contribute to another when added with it?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggb-new}{\code{ggb$new()}}
\item \href{#method-ggb-add_layers}{\code{ggb$add_layers()}}
\item \href{#method-ggb-add_slices}{\code{ggb$add_slices()}}
\item \href{#method-ggb-add_annotations}{\code{ggb$add_annotations()}}
\item \href{#method-ggb-add_image_labels}{\code{ggb$add_image_labels()}}
\item \href{#method-ggb-add_region_labels}{\code{ggb$add_region_labels()}}
\item \href{#method-ggb-render}{\code{ggb$render()}}
\item \href{#method-ggb-plot}{\code{ggb$plot()}}
\item \href{#method-ggb-clone}{\code{ggb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-new}{}}}
\subsection{Method \code{new()}}{
create a new ggb object. Note that inputs are always cloned to avoid
unintended modify-in-place behaviors of R6 classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$new(
  images = NULL,
  slices = NULL,
  layers = NULL,
  labels = NULL,
  action = NULL,
  annotations = NULL,
  region_labels = NULL,
  bg_color = "grey50",
  text_color = "black"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{images}}{a ggbrain_images object containing relevant images}

\item{\code{slices}}{a character vector of slices to extract}

\item{\code{layers}}{a list of ggbrain_layer objects}

\item{\code{action}}{the action to be taken when adding this object to an existing ggb}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-add_layers"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-add_layers}{}}}
\subsection{Method \code{add_layers()}}{
add layers from another ggb object to this one
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$add_layers(ilist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ilist}}{a list of ggbrain_layer objects. If a ggb object is passed, we
will get this list from obj$ggb_layers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-add_slices"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-add_slices}{}}}
\subsection{Method \code{add_slices()}}{
add slices to the existing vector of slices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$add_slices(slices = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slices}}{a character vector of slices to be appended to the existing slices}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-add_annotations"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-add_annotations}{}}}
\subsection{Method \code{add_annotations()}}{
add annotations to panels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$add_annotations(annotations = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annotations}}{a list or data.frame containing the annotations to add to each panel. Minimally,
the list or data.frame must contain \code{position} and \code{label} columns that define the position
and text to be added. Other arguments that pass through to ggplot2::annotate() can be provided as columns/elements
in \code{annotations} and these will be passed through to annotate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-add_image_labels"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-add_image_labels}{}}}
\subsection{Method \code{add_image_labels()}}{
add labels to a given image
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$add_image_labels(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{a named list of data.frame objects where the names denote corresponding images}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-add_region_labels"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-add_region_labels}{}}}
\subsection{Method \code{add_region_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$add_region_labels(labels = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-render"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-plot"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-plot}{}}}
\subsection{Method \code{plot()}}{
plot this ggb object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$plot()}\if{html}{\out{</div>}}
}

\subsection{Details}{
requires that required elements are in place already.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
