% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_layer.R
\name{ggbrain_layer}
\alias{ggbrain_layer}
\title{R6 class for a single layer of a ggbrain panel}
\description{
R6 class for a single layer of a ggbrain panel

R6 class for a single layer of a ggbrain panel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ggbrain_layer$new()}}
\item \href{#method-set_limits}{\code{ggbrain_layer$set_limits()}}
\item \href{#method-plot}{\code{ggbrain_layer$plot()}}
\item \href{#method-add_layer}{\code{ggbrain_layer$add_layer()}}
\item \href{#method-clone}{\code{ggbrain_layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a new ggbrain_layer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$new(
  layer_df = NULL,
  layer_scale = NULL,
  layer_limits = NULL,
  show_scale = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer_df}}{the data.frame containing image data for this layer. Must contain "dim1", "dim2",
and "value" as columns}

\item{\code{layer_scale}}{a ggplot scale object used for mapping the value column as the fill color for the
layer.}

\item{\code{layer_limits}}{if provided, sets the upper and lower bounds on the scale}

\item{\code{show_scale}}{if TRUE, show the scale on the plot legend}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_limits"></a>}}
\if{latex}{\out{\hypertarget{method-set_limits}{}}}
\subsection{Method \code{set_limits()}}{
set the limits for this layer's scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$set_limits(limits)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limits}}{a 2-element numeric vector setting the lower and upper limits on the layer's scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot this layer alone (mostly for debugging)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_layer"></a>}}
\if{latex}{\out{\hypertarget{method-add_layer}{}}}
\subsection{Method \code{add_layer()}}{
method to add this layer to an existing ggplot object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$add_layer(base_gg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_gg}}{the ggplot object to which we add the layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
