% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_plot.R
\name{ggbrain_plot}
\alias{ggbrain_plot}
\title{An R6 class for constructing a ggbrain plot from a ggbrain_slices object}
\description{
An R6 class for constructing a ggbrain plot from a ggbrain_slices object

An R6 class for constructing a ggbrain plot from a ggbrain_slices object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggbrain_plot-new}{\code{ggbrain_plot$new()}}
\item \href{#method-ggbrain_plot-generate_plot}{\code{ggbrain_plot$generate_plot()}}
\item \href{#method-ggbrain_plot-generate_panel}{\code{ggbrain_plot$generate_panel()}}
\item \href{#method-ggbrain_plot-clone}{\code{ggbrain_plot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_plot-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_plot-new}{}}}
\subsection{Method \code{new()}}{
instantiate a new instance of a ggbrain_plot object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_plot$new(slice_data, contrasts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slice_data}}{a ggbrain_slices object generated by ggbrain_images$get_slices()}

\item{\code{contrasts}}{a list or named character vector of contrasts to display}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_plot-generate_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_plot-generate_plot}{}}}
\subsection{Method \code{generate_plot()}}{
generate the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_plot$generate_plot(layers = NULL, slice_indices = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layers}}{a list of layers to be displayed on each panel, the order of which yields the
bottom-to-to drawing order within ggplot2. Each element of \code{layers} should be a list
that follows the approximate structure of the ggbrain_layer class, minimally including
the layer \code{name}, which is used to lookup data of images or contrasts within the
slice_data object.}

\item{\code{slice_indices}}{An optional subset of slice indices to display from the stored slice data}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
In addition to \code{name}, the elements of a layer can include
\code{layer_scale} a ggplot2 scale object for coloring the layer. Should be a scale_fill_* object.
\code{limits} the numeric limits to use for the color scale of this layer
\code{breaks} the scale breaks to use for the color scale of this layer
\code{show_scale} if FALSE, the color scale will not appear in the legend
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_plot-generate_panel"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_plot-generate_panel}{}}}
\subsection{Method \code{generate_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_plot$generate_panel(r)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
