% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBrainTheme.R
\name{ggBrainTheme}
\alias{ggBrainTheme}
\title{Set ggplot theme for ggBrain package}
\usage{
ggBrainTheme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Sets general size for text elements.}

\item{base_family}{Sets font familiy to use}
}
\value{
A ggplot Theme - list of class theme, gg.
}
\description{
Set ggplot theme for ggBrain package
}
\examples{
ggplot() + ggBrainTheme()
}
