% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_panel.R
\name{ggplot_add.ggbrain_panel}
\alias{ggplot_add.ggbrain_panel}
\title{S3 method to support adding ggbrain_layer objects to an existing ggplot object}
\usage{
\method{ggplot_add}{ggbrain_panel}(object, plot, object_name)
}
\description{
S3 method to support adding ggbrain_layer objects to an existing ggplot object
}
