% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_break_functions.R
\name{integer_breaks}
\alias{integer_breaks}
\title{breaks function to encourage integer-valued breaks, based on input from pretty}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{number of breaks (default = 5)}

\item{...}{Additional arguments passed to the pretty() function}
}
\description{
breaks function to encourage integer-valued breaks, based on input from pretty
}
\details{
Code from here: https://joshuacook.netlify.app/post/integer-values-ggplot-axis/
}
