% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integer_mode}
\alias{integer_mode}
\title{Finds the mode of an integer vector}
\arguments{
\item{v}{a vector of integers}

\item{demote_zeros}{if TRUE, zero will not be allowed to be the mode}
}
\value{
An integer representing the mode of the vector. If multiple modes are found, then the lowest-valued mode
is returned
}
\description{
Finds the mode of an integer vector
}
\keyword{internal}
