% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_slices.R
\name{is_cluster_slices_spec}
\alias{is_cluster_slices_spec}
\title{Check if an object is a cluster_slices_spec}
\usage{
is_cluster_slices_spec(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a cluster_slices_spec, FALSE otherwise
}
\description{
Check if an object is a cluster_slices_spec
}
\keyword{internal}
