% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat2df}
\alias{mat2df}
\title{Fast conversion of 2D mat to 3-column data.frame with dim1, dim2, value}
\arguments{
\item{mat}{A \code{matrix} to convert to data.frame}
}
\value{
A 3-column data.frame with dim1, dim2, and value
}
\description{
Converts a 2D numeric matrix into a 3-column data.frame
}
\details{
This function is a faster version of reshape2::melt for the simple 2-D case. It is about 2.5x faster than melt.
}
\author{
Michael Hallquist
}
\keyword{internal}
