% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{named_list}
\alias{named_list}
\title{little helper function to create named list from objects}
\usage{
named_list(...)
}
\arguments{
\item{...}{A set of arguments to be compiled into a list}
}
\description{
little helper function to create named list from objects
}
\details{
The names of the objects will form the names of the list elements
}
