% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPalette.R
\docType{data}
\name{palette_jet}
\alias{palette_jet}
\title{Colour Palettes common in MRI / PET Visualizations}
\format{
An object of class \code{character} of length 9.
}
\usage{
palette_jet
}
\description{
Colour Palettes common in MRI / PET Visualizations
}
\keyword{datasets}
