% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{plot.ggbrain_patchwork}
\alias{plot.ggbrain_patchwork}
\alias{print.ggbrain_patchwork}
\title{S3 method to allow for plot() syntax with rendered ggbrain patchwork objects}
\usage{
\method{plot}{ggbrain_patchwork}(x, ...)

\method{print}{ggbrain_patchwork}(x, ...)
}
\arguments{
\item{x}{the \code{ggbrain_patchwork} object to be plotted}

\item{...}{additional arguments. Not currently used}
}
\value{
\code{patchworkGrob} object, invisibly
}
\description{
S3 method to allow for plot() syntax with rendered ggbrain patchwork objects

default S3 method for ggbrain_patchwork objects (post-render)
}
