% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrainSlice.R
\name{plotBrainSlice}
\alias{plotBrainSlice}
\title{Set ggplot theme for ggBrain package}
\usage{
plotBrainSlice(
  image,
  mask = NULL,
  image_pixdim = NULL,
  mask_pixdim = NULL,
  ratio = 1,
  xDim = NULL,
  yDim = NULL,
  mask_min = NULL,
  mask_max = NULL,
  palette_mask = palette_heat,
  palette_image = palette_gray,
  alpha = 1
)
}
\arguments{
\item{base_size}{Sets general size for text elements.}

\item{base_family}{Sets font familiy to use}
}
\value{
A ggplot Theme - list of class theme, gg.
}
\description{
Set ggplot theme for ggBrain package
}
