% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{+.ggb}
\alias{+.ggb}
\title{addition operator for ggb object to support ggplot-like syntax}
\usage{
\method{+}{ggb}(o1, o2)
}
\arguments{
\item{o1}{the first object inheriting the ggb class}

\item{o2}{the second object inheriting the ggb class}
}
\value{
a modified version of the o1 object with o2 added to it
}
\description{
addition operator for ggb object to support ggplot-like syntax
}
\details{
Note that the addition operator always clones the underlying o1 object
rather than modifying it in place
}
