% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_images.R
\name{+.ggbrain_images}
\alias{+.ggbrain_images}
\title{addition operator for combining ggbrain_images objects}
\usage{
\method{+}{ggbrain_images}(o1, o2)
}
\arguments{
\item{o1}{first ggbrain_images object}

\item{o2}{second ggbrain_images object}
}
\value{
combined ggbrain_images object
}
\description{
addition operator for combining ggbrain_images objects
}
\details{
note that the addition does not modify either existing object. Rather,
the first object is cloned and the second is added to it. If you want to add one
ggbrain_images object to another in place (i.e., modifying the extant object), use
the $add() method.
}
