% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{render.ggb}
\alias{render.ggb}
\title{S3 method to allow for render(x) syntax with ggbrain (ggb) objects}
\usage{
\method{render}{ggb}(x, ...)
}
\arguments{
\item{x}{the \code{ggb} object to be rendered to a \code{ggbrain_patchwork} object}

\item{...}{additional arguments passed to the render method}
}
\value{
the ggbrain_patchwork object that can be handed off to patchwork and ggplot2
functions.
}
\description{
S3 method to allow for render(x) syntax with ggbrain (ggb) objects
}
