% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{add_images}
\alias{add_images}
\title{Add images to a ggb object}
\usage{
add_images(images = NULL)
}
\arguments{
\item{images}{a character vector or ggbrain_images object containing NIfTI images to add to this plot}
}
\value{
a ggb object with the relevant images and an action of 'add_images'
}
\description{
Add images to a ggb object
}
