% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_parser.R
\name{contrast_parser}
\alias{contrast_parser}
\title{helper function to calculate contrasts of one or more images using a combination of
image arithmetic and logical subsetting}
\usage{
contrast_parser(expr, data = NULL, default_val = NA_real_)
}
\arguments{
\item{expr}{a string or expression containing the image calculation to be performed}

\item{data}{a data.frame containing all variables used in \code{expr}. This will be used
to perform contrast calculations}

\item{default_val}{the value to be returned for any element of the contrast calculation that
does not pass through the arithmetic, either because of logical subsetting or because it
is exactly zero. In general, leave this as \code{NA_real_} unless you know what you're doing.}
}
\description{
helper function to calculate contrasts of one or more images using a combination of
image arithmetic and logical subsetting
}
\examples{
\dontrun{
  data <- matrix(rnorm(1000), ncol=5) \%>\% data.frame() \%>\%
    setNames(c("a", "f", "dsf", "abc", "b"))
  data$sdf <- rbinom(200, 1, .5)
 res <- contrast_parser("a + f + dsf [ sdf == 1 ]  * abc[ abc > 1 ] + b",data)
 res <- contrast_parser("dsf [ sdf == 1 ] + b", data)
 res <- contrast_parser("dsf * abc", data)
}
}
\author{
Michael Hallquist
}
\keyword{internal}
