% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{define}
\alias{define}
\title{Adds contrast definitions to the ggbrain plot}
\usage{
define(contrasts = NULL)
}
\description{
Adds contrast definitions to the ggbrain plot
}
\details{
Contrasts must take the form of \verb{<name> := <value expression>}.
Note that defining a contrast does not directly impact the appearance of the plot unless the
contrast is named in a geom_* layer.

Also note that contrasts can be specified in the definition of a layer. Thus, the \code{define} function
has two primary virtues. First, it allows for the conceptual separation of contrast definition versus usage
inside a geom_* layer, which is particularly useful if a contrast is used across several layers. Second, it allows
downstream layers to further modify the contrast, such as when we compute a
}
\examples{
\dontrun{
# simple example of a difference contrast, separating definition from usage in geom_brain
gg_obj <- ggbrain() +
  images(underlay = "T1.nii.gz", onset="zstat1.nii.gz", feedback="zstat2.nii.gz") +
  slices(c("x=25\%", x = "75\%")) +
  define("onset_gt_feedback := onset - feedback") +
  geom_brain("onset_gt_feedback")

# contrast definitions can also occur inline, yielding equivalent plots
gg_obj <- ggbrain() +
  images(underlay = "T1.nii.gz", onset="zstat1.nii.gz", feedback="zstat2.nii.gz") +
  slices(c("x=25\%", x = "75\%")) +
  geom_brain("onset - feedback")

# The use of contrasts() is helpful when layers modify the contrast (e.g., subsetting)
gg_obj <- ggbrain() +
  images(underlay = "T1.nii.gz", onset="zstat1.nii.gz", feedback="zstat2.nii.gz") +
  slices(c("x=25\%", x = "75\%")) +
  define("onset_gt_feedback := onset - feedback") +
  geom_brain("onset_gt_feedback[onset_gt_feedback > 0]", fill_scale=scale_fill_distiller("Pos diff", palette = "Reds")))

}


}
