% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{geom_brain}
\alias{geom_brain}
\title{Helper function to add a contrast layer to the plot. This is a thin wrapper around ggbrain_layer}
\usage{
geom_brain(...)
}
\arguments{
\item{name}{the name of this brain layer}

\item{definition}{a character string of the contrast or image definition used to define this layer.
Can be a simple image name (e.g., 'underlay') or a contrast string (e.g., \code{'overlay[overlay > 5]'})}

\item{fill_scale}{the color scale to be used for this layer (a scale_fill* object)}

\item{show_legend}{whether to show the color scale for this layer in the legend}
}
\value{
a ggb object populated with the ggb_layer and the action of 'add_layers'
}
\description{
Helper function to add a contrast layer to the plot. This is a thin wrapper around ggbrain_layer
}
\details{
Note that the fill_scale and limits must be specified at the time of the geom_contrast creation
in order for them to be mapped properly within ggplot. Because we overlay many raster layers in a ggplot
object that all use the fill aesthetic mapping, it becomes hard to map the color scales after the layer is
created using the typical + scale_fill_X syntax, and similarly for scale limits.
}
