% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{ggb}
\alias{ggb}
\title{Generic R6 base class that is inherited by all other ggbrain objects to support + semantics}
\description{
Generic R6 base class that is inherited by all other ggbrain objects to support + semantics

Generic R6 base class that is inherited by all other ggbrain objects to support + semantics
}
\details{
this object becomes a simple storage class that contains all relevant objects
(e.g., ggbrain_images) required to generate a brain plot
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ggb_images}}{ggbrain_images object for this plot}

\item{\code{ggb_slices}}{character string of slices to extract}

\item{\code{action}}{what should this ggb object contribute to another when added with it?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggb-new}{\code{ggb$new()}}
\item \href{#method-ggb-plot}{\code{ggb$plot()}}
\item \href{#method-ggb-clone}{\code{ggb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-new}{}}}
\subsection{Method \code{new()}}{
create a new ggb object. Note that inputs are always cloned to avoid
unintended modify-in-place behaviors of R6 classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$new(images = NULL, slices = NULL, action = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{images}}{a ggbrain_images object containing relevant images}

\item{\code{slices}}{a ggbrain_images object containing relevant images}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-plot"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-plot}{}}}
\subsection{Method \code{plot()}}{
plot this ggb object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$plot()}\if{html}{\out{</div>}}
}

\subsection{Details}{
requires that required elements are in place already.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
