% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain.R
\name{ggbrain}
\alias{ggbrain}
\title{Plot fMRI data on an underlay image}
\usage{
ggbrain(
  underlay = NULL,
  overlay = NULL,
  color_col = NULL,
  alpha_col = NULL,
  underlay_colorscale = scale_fill_gradient(low = "grey8", high = "grey92", na.value =
    "transparent"),
  negative_colorscale = scale_fill_distiller(palette = "Blues", direction = 1, na.value
    = "transparent"),
  positive_colorscale = scale_fill_distiller(palette = "Reds", na.value =
    "transparent"),
  slices = data.frame(coord = c("x = 25\%", "x = 50\%", "x = 75\%", "y = 25\%",
    "y = 50\%", "y = 75\%", "z = 25\%", "z = 50\%", "z = 75\%")),
  remove_null_space = TRUE,
  pos_thresh = 1,
  neg_thresh = -1,
  legend_label = expression(italic("z")),
  background_color = "gray10",
  text_color = "white",
  trim_underlay = c(0.01, 0.99),
  zero_underlay = 0.001,
  symmetric_legend = TRUE,
  panel_labels = NULL,
  underlay_contrast = "none",
  panel_borders = TRUE,
  theme_custom = NULL,
  base_size = 14
)
}
\arguments{
\item{underlay}{a 3D nifti image used for the image underlay (default b/w)}

\item{overlay}{a 4D nifti image used for plotting stats on underlay (color)}

\item{color_col}{a position in the 4th dim of overlay use to color plots}

\item{alpha_col}{a position in the 4th dim of overlay use to set alpha transparency of plots}

\item{underlay_colorscale}{A ggplot scale_fill_* function call used for coloration of underlay}

\item{remove_null_space}{If TRUE, quantiles are computed on the non-zero slices and plots are trimmed for
empty space.}

\item{pos_thresh}{The positive threshold to be applied to the \code{overlay} image. Any voxel that exceeds
the}

\item{trim_underlay}{Winsorize the extreme values of the underlay based on the low and high quantiles provided}

\item{zero_underlay}{Any voxels in the underlay image whose absolute values are less than \code{zero_underlay}
are set to precisely zero. This helps with small color variation in black/empty space.}

\item{overlay_colorscale}{A ggplot scale_fill_* function call used for coloration of overlay}
}
\description{
Plot fMRI data on an underlay image
}
