% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_label.R
\name{ggbrain_label}
\alias{ggbrain_label}
\title{R6 class for adding labels to a ggbrain_panel}
\description{
R6 class for adding labels to a ggbrain_panel

R6 class for adding labels to a ggbrain_panel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggbrain_label-new}{\code{ggbrain_label$new()}}
\item \href{#method-ggbrain_label-add_to_gg}{\code{ggbrain_label$add_to_gg()}}
\item \href{#method-ggbrain_label-clone}{\code{ggbrain_label$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_label-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_label-new}{}}}
\subsection{Method \code{new()}}{
create a new ggbrain_label object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_label$new(label_df = NULL, geom = "text", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_df}}{a data.frame containing labels to be printed on the panel. Must contain dim1, dim2, and label as columns.
The dim1 and dim2 columns control where the labels will appear on the panel}

\item{\code{geom}}{The geom type to be plotted. Must be "text" or "label", corresponding to geom_text and geom_label, respectively.}

\item{\code{...}}{All other arguments that will be passed directly to geom_text or geom_label such as hjust, size, and color}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_label-add_to_gg"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_label-add_to_gg}{}}}
\subsection{Method \code{add_to_gg()}}{
add this annotation to an existing ggplot object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_label$add_to_gg(base_gg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_gg}}{the ggplot object to which we add the layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_label-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_label-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_label$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
