% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_layer.R
\name{ggbrain_layer}
\alias{ggbrain_layer}
\title{R6 class for a single layer of a ggbrain panel}
\description{
R6 class for a single layer of a ggbrain panel

R6 class for a single layer of a ggbrain panel
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggbrain_layer-new}{\code{ggbrain_layer$new()}}
\item \href{#method-ggbrain_layer-set_limits}{\code{ggbrain_layer$set_limits()}}
\item \href{#method-ggbrain_layer-set_breaks}{\code{ggbrain_layer$set_breaks()}}
\item \href{#method-ggbrain_layer-set_name}{\code{ggbrain_layer$set_name()}}
\item \href{#method-ggbrain_layer-plot}{\code{ggbrain_layer$plot()}}
\item \href{#method-ggbrain_layer-add_to_gg}{\code{ggbrain_layer$add_to_gg()}}
\item \href{#method-ggbrain_layer-get_data}{\code{ggbrain_layer$get_data()}}
\item \href{#method-ggbrain_layer-get_name}{\code{ggbrain_layer$get_name()}}
\item \href{#method-ggbrain_layer-is_empty}{\code{ggbrain_layer$is_empty()}}
\item \href{#method-ggbrain_layer-clone}{\code{ggbrain_layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-new}{}}}
\subsection{Method \code{new()}}{
create a new ggbrain_layer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$new(
  name = NULL,
  data = NULL,
  layer_scale = NULL,
  limits = NULL,
  breaks = NULL,
  show_scale = TRUE,
  interpolate = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of this layer, used for referencing in layer and panel modifications}

\item{\code{data}}{the data.frame containing image data for this layer. Must contain "dim1", "dim2",
and "value" as columns}

\item{\code{layer_scale}}{a ggplot scale object used for mapping the value column as the fill color for the
layer.}

\item{\code{limits}}{if provided, sets the upper and lower bounds on the scale}

\item{\code{breaks}}{if provided, a function to draw the breaks on the color scale}

\item{\code{show_scale}}{if TRUE, show the scale on the plot legend}

\item{\code{interpolate}}{passes to geom_raster and controls whether the fill is interpolated over continuous space}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-set_limits"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-set_limits}{}}}
\subsection{Method \code{set_limits()}}{
set the limits for this layer's scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$set_limits(limits)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limits}}{a 2-element numeric vector setting the lower and upper limits on the layer's scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-set_breaks"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-set_breaks}{}}}
\subsection{Method \code{set_breaks()}}{
set the breaks element of this layer's scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$set_breaks(breaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breaks}}{a function used to label the breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-set_name}{}}}
\subsection{Method \code{set_name()}}{
set the layer name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a character string defining the layer's name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-plot"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-plot}{}}}
\subsection{Method \code{plot()}}{
plot this layer alone (mostly for debugging)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-add_to_gg"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-add_to_gg}{}}}
\subsection{Method \code{add_to_gg()}}{
method to add this layer to an existing ggplot object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$add_to_gg(base_gg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_gg}}{the ggplot object to which we add the layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-get_data}{}}}
\subsection{Method \code{get_data()}}{
return the data.frame associated with this layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$get_data(add_layer_name = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_layer_name}}{if TRUE, adds a \code{layer_name} column to the data.frame for record-keeping.
Default: FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-get_name}{}}}
\subsection{Method \code{get_name()}}{
return the layer name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
returns TRUE if all values are NA or if the data has 0 rows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_layer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_layer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
