% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_slices.R
\name{ggbrain_slices}
\alias{ggbrain_slices}
\title{R6 class for managing slice data for ggbrain plots}
\description{
R6 class for managing slice data for ggbrain plots

R6 class for managing slice data for ggbrain plots
}
\examples{

## ------------------------------------------------
## Method `ggbrain_slices$compute_contrasts`
## ------------------------------------------------

\dontrun{
  slc <- ggbrain_slices$new(slice_df=my_data)
  slc$compute_contrasts(list(pos_vals="overlay[overlay> 0]"))
}
}
\author{
Michael Hallquist
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{slice_index}}{read-only access to the slice_index containing the slice numbers}

\item{\code{coord_input}}{the input string used to lookup the slices}

\item{\code{coord_label}}{the calculated x, y, or z coordinate of the relevant slice}

\item{\code{slice_number}}{the slice number along the relevant axis of the 3D image matrix}

\item{\code{slice_data}}{a nested list of data.frames where each element contains all data relevant to
that slice and the list elements within are each a given image}

\item{\code{slice_matrix}}{the slice data in matrix form}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ggbrain_slices-new}{\code{ggbrain_slices$new()}}
\item \href{#method-ggbrain_slices-compute_contrasts}{\code{ggbrain_slices$compute_contrasts()}}
\item \href{#method-ggbrain_slices-as_tibble}{\code{ggbrain_slices$as_tibble()}}
\item \href{#method-ggbrain_slices-get_image_names}{\code{ggbrain_slices$get_image_names()}}
\item \href{#method-ggbrain_slices-clone}{\code{ggbrain_slices$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_slices-new"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_slices-new}{}}}
\subsection{Method \code{new()}}{
create a ggbrain_slices object based
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_slices$new(slice_df = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slice_df}}{a data.frame generated by ggbrain_images$get_slices()}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If this becomes a user-facing/exported class, we may want a more friendly constructor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_slices-compute_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_slices-compute_contrasts}{}}}
\subsection{Method \code{compute_contrasts()}}{
computes contrasts of the sliced image data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_slices$compute_contrasts(contrast_list = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contrast_list}}{a named list or character vector containing contrasts to be computed.
The names of the list form the contrast names, while the values should be character strings
that use standard R syntax for logical tests, subsetting, and arithmetic}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  slc <- ggbrain_slices$new(slice_df=my_data)
  slc$compute_contrasts(list(pos_vals="overlay[overlay> 0]"))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_slices-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_slices-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
convert the slices object into a data.frame with list-columns for slice data elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_slices$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_slices-get_image_names"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_slices-get_image_names}{}}}
\subsection{Method \code{get_image_names()}}{
returns a vector of the names of all image and contrast data available in this
ggbrain_slices object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_slices$get_image_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggbrain_slices-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ggbrain_slices-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ggbrain_slices$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
