% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_label.R
\name{ggplot_add.ggbrain_label}
\alias{ggplot_add.ggbrain_label}
\title{S3 method to support adding ggbrain_label objects to an existing ggplot object}
\usage{
\method{ggplot_add}{ggbrain_label}(object, plot, object_name)
}
\description{
S3 method to support adding ggbrain_label objects to an existing ggplot object
}
