% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_layer.R
\name{ggplot_add.ggbrain_layer}
\alias{ggplot_add.ggbrain_layer}
\title{S3 method to support adding ggbrain_layer objects to an existing ggplot object}
\usage{
\method{ggplot_add}{ggbrain_layer}(object, plot, object_name, ...)
}
\arguments{
\item{object}{the ggbrain_layer object to be added to an existing ggplot}

\item{plot}{the ggplot object}

\item{object_name}{not used, but required by ggplot_add}

\item{...}{additional arguments, not currently used}
}
\description{
S3 method to support adding ggbrain_layer objects to an existing ggplot object
}
