% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{montage}
\alias{montage}
\title{Convenience function to add many slices in a montage along one of the 3D planes}
\usage{
montage(plane = NULL, n = 12, min = 0.1, max = 0.9)
}
\arguments{
\item{plane}{a character string specifying the 3D plane: "sagittal", "axial", "coronal", "x", "y", or "z"}

\item{n}{number of slices to add in this plane. Default: 12}

\item{min}{the lowest quantile to be included in the montage (between 0 and 1). Default: 0.1}

\item{max}{the highest quantilt to be included in the montage (between 0 and 1). Default: 0.9}
}
\description{
Convenience function to add many slices in a montage along one of the 3D planes
}
\details{
this can be used with \code{add_slices} to make a quick montage, such as \verb{add_slices(montage("axial", 10)}.
}
