% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{plot.ggb}
\alias{plot.ggb}
\title{S3 method to allow for plot() syntax with ggbrain (ggb) objects}
\usage{
\method{plot}{ggb}(x, ...)
}
\arguments{
\item{x}{the \code{ggb} object to be plotted}

\item{...}{additional arguments passed to the plot method}
}
\value{
NULL, invisibly
}
\description{
S3 method to allow for plot() syntax with ggbrain (ggb) objects
}
