% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain_panel.R
\name{plot.ggbrain_panel}
\alias{plot.ggbrain_panel}
\title{S3 method to allow for plot() syntax with ggbrain_panel objects}
\usage{
\method{plot}{ggbrain_panel}(object)
}
\description{
S3 method to allow for plot() syntax with ggbrain_panel objects
}
