% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_break_functions.R
\name{range_breaks}
\alias{range_breaks}
\title{breaks function for including min + max with labels, and a few unlabeled ticks in between}
\usage{
range_breaks(n = 3, digits = 2)
}
\arguments{
\item{n}{number of breaks added within the min-max range}

\item{digits}{number of decimal places to display}
}
\description{
breaks function for including min + max with labels, and a few unlabeled ticks in between
}
