% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{slices}
\alias{slices}
\title{Function for adding a single slice to the plot, allowing for additional panel attributes to be passed through}
\usage{
slices(
  coordinates = NULL,
  title = NULL,
  bg_color = NULL,
  text_color = NULL,
  border_color = NULL,
  border_size = NULL,
  xlab = NULL,
  ylab = NULL,
  theme_custom = NULL
)
}
\arguments{
\item{coordinates}{a string specifying the x, y, or z coordinate of the slice to be added. Follows the same logic as add_slices.}

\item{title}{a title for the panel added to the ggplot object using ggtitle()}

\item{bg_color}{the color used for the background of the panel. Default: \code{'gray10'} (nearly black)}

\item{text_color}{the color used for text displayed on the panel. Default: \code{'white'}.}

\item{border_color}{the color used for drawing a border around on the panel. Default: \code{'gray50'}
(though borders are not drawn by default).}

\item{border_size}{the size of the border line drawn around the panel. Default: NULL. If this value is
greater than zero, a border of this size and with color \code{border_color} will be drawn around the panel}

\item{xlab}{The label to place on x axis. Default is NULL.}

\item{ylab}{The label to place on y axis. Default is NULL.}

\item{theme_custom}{Any custom theme() settings to be added to the panel}
}
\description{
Function for adding a single slice to the plot, allowing for additional panel attributes to be passed through
}
\details{
note that if you pass in multiple coordinates (as a vector), the title, bg_color, and other attributes are all treated
as the same for slices added by this operation. Thus, if you want to customize specific slices or groups of slices, use
multiple addition operations, as in \code{slices(c('x=10', 'y=15'), bg_color='white') + slices(c('x=18', 'y=22'), bg_color='black')}.
}
